% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elementsAllEqual}
\alias{elementsAllEqual}
\title{Determine if a vector has identical elements}
\usage{
elementsAllEqual(X, Y = 1e-06)
}
\arguments{
\item{X}{Numerical vector}

\item{Y}{Tolerance for pairwise comparison Default is 1e-6.}
}
\description{
Determine if a vector has identical elements
}
