% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptDecrypt.R
\name{encrypt}
\alias{encrypt}
\alias{decrypt}
\title{Scramble/Unscramble Strings}
\usage{
encrypt(STR, base64 = TRUE)

decrypt(STR, base64 = TRUE)
}
\arguments{
\item{STR}{Character vector to be encrypted/decrypted.}

\item{base64}{Whether or not to use base64 encoding/decoding. Default is TRUE.}
}
\description{
Intended to protect sensitive information from human eyes (but not from serious attacks).
}
