% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{getGHUserList}
\alias{getGHUserList}
\title{Get List of Github User}
\usage{
getGHUserList(FILE)
}
\arguments{
\item{FILE}{CSV file containing a list of user name and associated email name.}
}
\value{
User list as `data.table`
}
\description{
Get List of Github User
}
